python/consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.32
protobuf==3.19.6
PyMySQL==1.0.2
